1) Data Cleaning and Analysis
This chapter covers Data Aggregation, Combining Data Using Pandas, Transforming Data with Pandas, Working with Strings in Pandas and Working with Missing and Duplicate Data. 
This project is about Cleaning and Analyzing exit surveys of employees from the two educational institutes in Queensland, Australia. In this project we have performed cleaning, combining and analysis of two datasets belonging to these two institutions.